feat: Initial UI commit
This commit sets up the MudBlazor framework for the frontend and
adds a basic page layout for both the desktop and mobile site.

The implementation has diverged from the Figma design due to the
differences in the components available, but the experience is
kept as near as possible to the concept.