chore(ui): Move everything possible to WASM
This commit moves most of the UI to WebAssembly client side
rendering. This allows better integration with the host environment.

Some features are lost, such as the template account management
pages provided by Identity Framework, but we would need to
re-write those anyway to match with our theme. These pages will be
re-implemented as one Client page connecting to a controller in the
future.