fix(ui): Send local date time in WizardPageBase
Stuff is not failing on the local development server since it has
the same time zone as the Sydney airport. However, deployment
servers may be in different timezones and reject dates with
unspecified times.

We therefore standardize everything in the WizardPageBase to be the
local time of the corresponding airports.